requirements.txt
autogen==0.2.0
PyYAML==6.0
urllib3==1.26.15
requests==2.28.2
google-generativeai==0.3.1
anthropic==0.5.0
openai==0.27.8